api/requirements.txt
Flask
Flask-SQLAlchemy
Flask-CORS
argon2-cffi
psycopg2-binary
PyJWT